Removed undefined epoch macro references
This package currently has no epoch, so it doesn't need an epoch
in self references. The macro wasn't defined and resulted in
broken requires definitions. A more complicated epoch reference
use could have been done to handle the no epoch case, but it seemed
safer to just remove it to get the package working again.